rename: choose-template.component --> choose-template.dialog
