+++ src/app/dialogs/choose-template/choose-template.dialog.html
